openai_agents_sdk_with_fast_api/pyproject.toml
[project]
name = "openai-agents-sdk-with-fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "openai-agents[litellm]>=0.0.17",
    "python-dotenv>=1.1.0",
]
